[PAGOPA-1364] test: refactoring and health check
