+++ integration-test/src/features/file-upload.feature
@@ -1,10 +1,10 @@
 Feature: file handling: { POST }
 
   Background:
-    """Given info ok"""
+    Given GPD-Upload running
 
   Scenario: File upload
     Given zip file of 100 payment-position
-    When the client send POST to /brokers/btest/organizations/ectest/debtpositions/file
+    When the client send POST to /brokers/brokertest/organizations/ectest/debtpositions/file
     Then check statusCode is 202
     And check location header regex ^brokers/(.+)/organizations/(.+)/debtpositions/file/(.+)/status$
